input button works in terminal
