+++ Functions_for_user_data.py
@@ -1,21 +1,10 @@
 from datetime import datetime
-import matplotlib.pyplot as plt
-import pandas as pd
 
-def __init__():
-    print("Welcome! May I please have your username?")
-    username = input("Username: ")
-    username = username.lower()
-    user_data = username + ".csv"
 
-    try:
-        open(user_data, "r")
-    except:
-        with open(user_data, "a") as new_file:
-            new_file.write("date, food, water, sleep, stress, activity\n")
 # TODO - 
 
-def query_user_data() -> list:
+def query_user_data():
+    # TODO - make this make an entry screen
     # taking in user data of 5 questions -
     # food, water, sleep, stress, physical activity
     # on a scale of 1 to 10
@@ -53,9 +42,3 @@ def write_user_data(new_line: list, file_name: str):
         line = line[:-1] + "\n"
         my_file.write(line)
     print("File saved!")
-
-def plot_by_time():
-    time_series = pd.read_csv(user_data) # reload the saved data as a dataframe for analysis
-    print(time_series)
-    time_series.plot() #make a quick plot to view the data
-    plt.show()
