+++ main.py
@@ -1,5 +1,6 @@
 import Functions_for_user_data as fud
-
+import pandas as pd
+import matplotlib.pyplot as plt
 from tkcalendar import Calendar #, DateEntry
 
 try:
@@ -10,6 +11,17 @@ except ImportError:
     import ttk
 
 # TODO - DT -> Pandas, QT -> TK
+
+print("Welcome! May I please have your username?")
+username = input("Username: ")
+username = username.lower()
+user_data = username + ".csv"
+
+try:
+    open(user_data, "r")
+except:
+    with open(user_data, "a") as new_file:
+        new_file.write("date, food, water, sleep, stress, activity\n")
 
 
 #new_data = fud.query_user_data()
@@ -31,6 +43,15 @@ def show_example_calendar():
     cal.pack(fill="both", expand=True)
     ttk.Label(top, text="Hover over the events.").pack()
 
+def plot_by_time():
+    time_series = pd.read_csv(user_data) # reload the saved data as a dataframe for analysis
+    print(time_series)
+    time_series.plot() #make a quick plot to view the data
+    plt.show()
+
+def input_data():
+    fud.write_user_data(fud.query_user_data(), user_data)
+    
 
 root = tk.Tk()
 root.title("DAT Biotracker")
@@ -41,8 +62,8 @@ root.rowconfigure(0, weight = 1)
 
 ttk.Label(mainframe, text = "Welcome").grid(column = 2, row = 1, sticky = tk.N)
 
-ttk.Button(mainframe, text = "Input", command = fud.query_user_data).grid(column = 1, row = 3, sticky = tk.S)
-ttk.Button(mainframe, text = "Output", command = fud.plot_by_time).grid(column = 2, row = 3, sticky = tk.S)
+ttk.Button(mainframe, text = "Input", command = input_data).grid(column = 1, row = 3, sticky = tk.S)
+ttk.Button(mainframe, text = "Output", command = plot_by_time).grid(column = 2, row = 3, sticky = tk.S)
 ttk.Button(mainframe, text='Example calendar with events', command = show_example_calendar).grid(column = 3, row = 3, sticky = tk.S)
 
 root.mainloop()
